pyproject.toml
[tool.poetry]
name = "acolyte"
version = "1.0.0"
description = "The Acolyte discord bot"
authors = ["mari <mari@hakke.ro>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^1.7.3"
python-ffmpeg = "^1.0.11"
aiofiles = "^0.7.0"
python-dotenv = "^0.18.0"
watchdog = "^2.1.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
